+++ CHANGELOG.md
@@ -24,6 +24,7 @@ You can find out backwards-compatibility policy [here](https://github.com/hynek/
 - Support for intra-payload headers like the ones used in OpenPGP keys using the `meta_headers` property.
   [#75](https://github.com/hynek/pem/pull/75)
 - `pem.parse_file()` now accepts also [`pathlib.Path`](https://docs.python.org/3/library/pathlib.html#pathlib.Path) objects.
+- `pem.parse()` now also accepts `str`.
 - Added `text_payload`, `bytes_payload` and `decoded_payload` properties to all PEM objects that allow to directly access the payload without the envelope and possible headers.
   [#74](https://github.com/hynek/pem/pull/74)